// Frontend/src/models/ShoesModel.ts
class ShoesModel {
    shoesId: number = 0;
    categoryId: number = 0;
    title: string = '';
    description: string = '';
    price: number = 0;
    bought: number = 0;
    stock: number = 0;
    imageName: string = '';
    imageLink: string = '';
    shoppingBasket: number = 0;
    favorite: number = 0;
    imageNameFront: string = '';
    imageNameAbove: string = '';
    imageNameBack: string = '';
    imageNameDown: string = '';
    video: string = '';
    color: string = '';
}

export default ShoesModel;